refactor: Move BetaTesterFactory helper to common/
