+++ lms/djangoapps/courseware/tests/test_self_paced_overrides.py
@@ -6,8 +6,8 @@ from unittest.mock import patch
 import pytz
 from django.test.utils import override_settings
 
+from common.djangoapps.student.tests.factories import BetaTesterFactory
 from lms.djangoapps.courseware.access import has_access
-from lms.djangoapps.courseware.tests.factories import BetaTesterFactory
 from lms.djangoapps.ccx.tests.test_overrides import inject_field_overrides
 from lms.djangoapps.courseware.field_overrides import OverrideFieldData, OverrideModulestoreFieldData
 from lms.djangoapps.discussion.django_comment_client.utils import get_accessible_discussion_xblocks
