+++ lms/djangoapps/courseware/tests/test_view_authentication.py
@@ -9,12 +9,12 @@ from unittest.mock import patch
 import pytz
 from django.urls import reverse
 
+from common.djangoapps.student.tests.factories import BetaTesterFactory
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
 from common.djangoapps.student.tests.factories import InstructorFactory
 from common.djangoapps.student.tests.factories import StaffFactory
 from lms.djangoapps.courseware.access import has_access
 from lms.djangoapps.courseware.tests.factories import (
-    BetaTesterFactory,
     OrgInstructorFactory,
     OrgStaffFactory,
 )
